alteracao de padrao de apstas e criacao de herança para as classes de interface da api
