+++ src/dependencies/web_scraping.py
