+++ src/routes/importacao.py
@@ -1,18 +1,19 @@
-from api.schemas import models_db as models
 import os
+
+from src.models import models_db as models
 
 from fastapi import APIRouter, status, Depends, HTTPException
 from sqlalchemy.orm import Session
 from typing import Annotated
 
-from api.schemas.models_db import Exportacao
-from api.dependencies.database import SessionLocal
-from api.schemas.models_api import ExportacaoBase
-from api.services.authentication import get_current_user
+from src.models.models_db import Importacao
+from src.dependencies.database import SessionLocal
+from src.models.api.model_importacao_api import ImportacaoBase, ImportacaoInsert
+from src.services.authentication import get_current_user
 
 
 router = APIRouter(
-    tags=['Exportacao'],
+    tags=['Importacao'],
     dependencies=[Depends(get_current_user)],
     responses={401: {'detail': os.environ.get('ERRO_401')}}
 )
@@ -30,16 +31,16 @@ db_dependency = Annotated[Session, Depends(get_db)]
 user_dependency = Annotated[Session, Depends(get_current_user)]
 
 
-@router.get('/exportacao/{id_exportacao}', status_code=status.HTTP_200_OK)
-async def exportacao_id(
-        id_exportacao: int,
+@router.get('/importacao/{id_importacao}', status_code=status.HTTP_200_OK)
+async def importacao_id(
+        id_importacao: int,
         db: db_dependency
 ):
     """
-    Obtém um item da tabela de exportacao pelo ID.
+    Obtém um item da tabela de importacao pelo ID.
 
     Args:
-        id_exportacao (int): O ID da exportacao.
+        id_importacao (int): O ID da importacao.
         db: Sessão do banco de dados.
 
     Returns:
@@ -47,26 +48,26 @@ async def exportacao_id(
             ou gera HTTP_404_NOT_FOUND se não encontrado.
     """
 
-    exportacao = db.query(models.Exportacao).filter(models.Exportacao.id == id_exportacao).first()
+    importacao = db.query(models.Importacao).filter(models.Importacao.id == id_importacao).first()
 
-    if exportacao is None:
+    if importacao is None:
         raise HTTPException(status_code=404, detail=os.environ.get('ERRO_404'))
 
-    return exportacao
+    return importacao
 
 
-@router.get('/exportacao', status_code=status.HTTP_200_OK)
-async def total_exportacao(
+@router.get('/importacao', status_code=status.HTTP_200_OK)
+async def total_importacao(
         db: db_dependency
 ):
     """
-    Obtém todos os dados da tabela de exportacao.
+    Obtém todos os dados da tabela de importacao.
 
     Args:
         db: Sessão do banco de dados.
 
     Returns:
-        list[Exportacao]: Uma lista de objetos Exportacao.
+        list[Importacao]: Uma lista de objetos Importacao.
 
     Raises:
         HTTPException: Com status code 500 se houver um erro ao obter os dados.
@@ -74,23 +75,24 @@ async def total_exportacao(
 
     try:
         # retorna todas as linhas da tabela
-        return db.query(models.Exportacao).all()
+        return db.query(models.Importacao).all()
     except Exception as e:
         print(e)
         raise HTTPException(status_code=500, detail="Erro ao obter os dados da tabela")
 
 
-@router.post('/exportacao/filtragem')
-async def filtrar_exportacao(
-        exportacao: ExportacaoBase,
+@router.post('/importacao/filtragem')
+async def filtrar_importacao(
+        importacao: ImportacaoBase,
         db: db_dependency
 ):
+
     """
-    Filtra dados da tabela de exportacao com base nos critérios fornecidos.
+    Filtra dados da tabela de importacao com base nos critérios fornecidos.
     Necessário passar pelo menos um dos parâmetros para retornar algo.
 
     Args:
-        exportacao (ExportacaoBase): Objeto com os critérios de filtro.
+        importacao (ImportacaoBase): Objeto com os critérios de filtro.
             Os campos disponíveis para filtro são:
             * id (int, optional): ID único da entrada.
             * categoria (str, optional): Categoria do produto.
@@ -101,38 +103,38 @@ async def filtrar_exportacao(
         db: Sessão do banco de dados.
 
     Returns:
-        list[Exportacao]: Uma lista de objetos Exportacao que correspondem aos filtros.
+        list[Importacao]: Uma lista de objetos Importacao que correspondem aos filtros.
 
     Raises:
         HTTPException: Com status code 500 se houver um erro ao filtrar os dados.
     """
 
     try:
-        query = db.query(models.Exportacao)
+        query = db.query(models.Importacao)
 
         # Adiciona filtro pelo ID se fornecido
-        if exportacao.id is not None:
-            query = query.filter(models.Exportacao.id == exportacao.id)
+        if importacao.id is not None:
+            query = query.filter(models.Importacao.id == importacao.id)
 
         # Adiciona filtro pela categoria se fornecida
-        if exportacao.categoria is not None:
-            query = query.filter(models.Exportacao.categoria == exportacao.categoria)
+        if importacao.categoria is not None:
+            query = query.filter(models.Importacao.categoria == importacao.categoria)
 
         # Adiciona filtro pelo nome se fornecido
-        if exportacao.nome is not None:
-            query = query.filter(models.Exportacao.nome == exportacao.nome)
+        if importacao.nome is not None:
+            query = query.filter(models.Importacao.nome == importacao.nome)
 
         # Adiciona filtro pelo ano se fornecido
-        if exportacao.ano is not None:
-            query = query.filter(models.Exportacao.ano == exportacao.ano)
+        if importacao.ano is not None:
+            query = query.filter(models.Importacao.ano == importacao.ano)
 
         # Adiciona filtro pelo valor de produção se fornecido
-        if exportacao.quantidade is not None:
-            query = query.filter(models.Exportacao.quantidade == exportacao.quantidade)
+        if importacao.quantidade is not None:
+            query = query.filter(models.Importacao.quantidade == importacao.quantidade)
 
         # Adiciona filtro pelo valor de produção se fornecido
-        if exportacao.valor is not None:
-            query = query.filter(models.Exportacao.valor == exportacao.valor)
+        if importacao.valor is not None:
+            query = query.filter(models.Importacao.valor == importacao.valor)
 
         # Executa a consulta e retorna os resultados
         return query.all()
@@ -142,18 +144,17 @@ async def filtrar_exportacao(
         raise HTTPException(status_code=500, detail="Erro ao filtrar os dados de produção")
 
 
-@router.post('/exportacao', status_code=status.HTTP_201_CREATED)
-async def insere_exportacao(
-        comercializacao: ExportacaoBase,
+@router.post('/importacao', status_code=status.HTTP_201_CREATED)
+async def insere_importacao(
+        comercializacao: ImportacaoInsert,
         db: db_dependency
 ):
-
     """
-    Insere dados da tabela de exportacao com base nos critérios fornecidos.
+    Insere dados da tabela de importacao com base nos critérios fornecidos.
     Necessário passar pelo menos um dos parâmetros para retornar algo.
 
     Args:
-        exportacao (ExportacaoBase): Objeto com os critérios de filtro.
+        importacao (ImportacaoBase): Objeto com os critérios de filtro.
             Os campos disponíveis para filtro são:
             * categoria (str, optional): Categoria do produto.
             * nome (str, optional): Nome do produto.
@@ -163,14 +164,13 @@ async def insere_exportacao(
         db: Sessão do banco de dados.
 
     Returns:
-        list[Exportacao]: Uma lista de objetos Exportacao que correspondem aos filtros.
+        list[Importacao]: Uma lista de objetos Importacao que correspondem aos filtros.
 
     Raises:
         HTTPException: Com status code 500 se houver um erro ao filtrar os dados.
     """
 
-
-    create_exportacao_model = Exportacao(
+    create_importacao_model = Importacao(
         categoria=comercializacao.categoria,
         nome=comercializacao.nome,
         ano=comercializacao.ano,
@@ -178,26 +178,26 @@ async def insere_exportacao(
         valor=comercializacao.valor
     )
 
-    db.add(create_exportacao_model)
+    db.add(create_importacao_model)
     db.commit()
 
-    db.refresh(create_exportacao_model)
+    db.refresh(create_importacao_model)
 
-    return {'id': create_exportacao_model.id}
+    return {'id': create_importacao_model.id}
 
 
-@router.put('/exportacao/{id_exportacao}', status_code=status.HTTP_204_NO_CONTENT)
-async def altera_exportacao(
-        id_exportacao: int,
-        exportacao: ExportacaoBase,
+@router.put('/importacao/{id_importacao}', status_code=status.HTTP_204_NO_CONTENT)
+async def altera_importacao(
+        id_importacao: int,
+        importacao: ImportacaoInsert,
         db: db_dependency
 ):
     """
-    Altera os dados referentes a um item existente na tabela de exportacao.
+    Altera os dados referentes a um item existente na tabela de importacao.
 
     Args:
-        id_exportacao (int): O ID da exportacao a ser alterada.
-        exportacao (ExportacaoBase): Objeto com os critérios de filtro.
+        id_importacao (int): O ID da importacao a ser alterada.
+        importacao (ImportacaoBase): Objeto com os critérios de filtro.
             Os campos disponíveis para filtro são:
             * categoria (str, optional): Categoria do produto.
             * nome (str, optional): Nome do produto.
@@ -207,23 +207,23 @@ async def altera_exportacao(
         db: Sessão do banco de dados.
 
     Returns:
-        list[Exportacao]: Uma lista de objetos Exportacao que correspondem aos filtros.
+        list[Importacao]: Uma lista de objetos Importacao que correspondem aos filtros.
 
     Raises:
         HTTPException: Com status code 500 se houver um erro ao filtrar os dados.
     """
 
     # Busca o item no banco de dados pelo ID
-    exportacao_model = db.query(Exportacao).filter(Exportacao.id == id_exportacao).first()
-    if exportacao_model is None:
+    importacao_model = db.query(Importacao).filter(Importacao.id == id_importacao).first()
+    if importacao_model is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=os.environ.get('ERRO_404'))
 
     # Atualiza os campos do objeto com os novos valores
-    exportacao_model.categoria = exportacao.categoria
-    exportacao_model.nome = exportacao.nome
-    exportacao_model.ano = exportacao.ano
-    exportacao_model.quantidade = exportacao.quantidade
-    exportacao_model.valor = exportacao.valor
+    importacao_model.categoria = importacao.categoria
+    importacao_model.nome = importacao.nome
+    importacao_model.ano = importacao.ano
+    importacao_model.quantidade = importacao.quantidade
+    importacao_model.valor = importacao.valor
 
     # Realiza o commit para persistir as alterações no banco de dados
     db.commit()
@@ -232,27 +232,27 @@ async def altera_exportacao(
     return None
 
 
-@router.delete('/exportacao/{id_exportacao}', status_code=status.HTTP_204_NO_CONTENT)
-async def deleta_exportacao(
-        id_exportacao: int,
+@router.delete('/importacao/{id_importacao}', status_code=status.HTTP_204_NO_CONTENT)
+async def deleta_importacao(
+        id_importacao: int,
         db: db_dependency
 ):
     """
-    Deleta um item da tabela de exportacao pelo ID.
+    Deleta um item da tabela de importacao pelo ID.
 
     Args:
-        id_exportacao (int): O ID da exportacao a ser deletada.
+        id_importacao (int): O ID da importacao a ser deletada.
         db: Sessão do banco de dados.
 
     Returns:
         None: Retorna um status HTTP 204 No Content em caso de sucesso.
 
     Raises:
-        HTTPException: Com status code 404 Not Found se a exportacao não for encontrada.
+        HTTPException: Com status code 404 Not Found se a importacao não for encontrada.
     """
 
     # Busca o item no banco de dados pelo ID
-    comercializacao_model = db.query(Exportacao).filter(Exportacao.id == id_exportacao).first()
+    comercializacao_model = db.query(Importacao).filter(Importacao.id == id_importacao).first()
     if comercializacao_model is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=os.environ.get('ERRO_404'))
