+++ src/services/tratamento_dados_tabela.py
